separate model and component classes
