+++ src/component/Inspector.scss
