+++ src/component/InspectorSearchField.js
