+++ src/index.js
@@ -1,3 +1,3 @@
-import Inspector from "./Inspector";
+import Inspector from "./component/Inspector";
 
 export default Inspector;
