+++ src/model/RefScope.js
