+++ src/model/searchUtils.js
