+++ test/component/InspectorSearchField.test.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
-import InspectorSearchField from "../src/InspectorSearchField";
+
+import InspectorSearchField from "../../src/component/InspectorSearchField";
 
 describe("renders correctly", () => {
     it("with minimal/default props", () => {
